api hooks: minor refacoring
