+++ packages/admin/src/tenant/lib/index.ts
@@ -1,6 +1,8 @@
 export * from './auth'
-export * from './facade'
-export * from './mutation'
-export * from './query'
+export * from './useSingleTenantMutation'
+export * from './useMutation'
+export * from './useQuery'
 export * from './requestState'
 export * from './useForm'
+export { GQLVariable } from './variables'
+export type { GQLVariableType } from './variables'
