+++ packages/admin/src/tenant/mutations/otp/disableOtp.ts
@@ -1,4 +1,4 @@
-import { GQLVariable, useSingleTenantMutation } from '../../lib/facade'
+import { GQLVariable, useSingleTenantMutation } from '../../lib'
 
 const DisableOtpMutation = `
 disableOtp {
